perf: halve lane position column to ushort
Seek(Span<ushort>, bool) replaces the uint column: 2 B/row positions,
32-lane vector run scan, duration capped at 65535 at bind. Interleaved
A/B vs ff16190 (controls flat): uniform 0.19->0.15-0.16, backward
0.19->0.16, waves 0.21-0.22->0.19-0.20 ns/row; staggered singles
2.79-2.92->3.13, catch-up 5.66-5.80->6.23 (+8%, 16-bit scalar ops in
the singleton path). Kept: crowd shapes are the lane's purpose and the
halved column halves every host position-cache budget. Public API
approval updated; docs carry the A/B verdict.

+++ benchmarks/Alpha/DataAuthored.cs
@@ -214,7 +214,7 @@ internal sealed class LaneCase : IDisposable
     internal const int Operations = 4096;
     private readonly int[] _deltas = new int[Operations];
     private readonly TimelineAsset _asset;
-    private readonly uint[] _positions = new uint[1];
+    private readonly ushort[] _positions = new ushort[1];
     private readonly float[] _values = new float[1];
     private readonly long[] _cycles = new long[1];
 
@@ -240,7 +240,7 @@ internal sealed class LaneCase : IDisposable
 
     public void Dispose() => _asset.Dispose();
 
-    internal static long Checksum(uint position, long cycle, float value)
+    internal static long Checksum(ushort position, long cycle, float value)
         => unchecked((long)position * 31 + cycle * 7 + (long)value);
 
     internal static byte[] Bake(TimelineShape shape)
@@ -296,6 +296,6 @@ internal sealed class LaneCase : IDisposable
             position = next.Position;
             cycle = next.Cycle;
         }
-        return Checksum(position, cycle, value);
+        return Checksum((ushort)position, cycle, value);
     }
 }
